+++ test/Confluent.Kafka.IntegrationTests/Tests/Producer_DisableDeliveryReports.cs
@@ -51,25 +51,25 @@ namespace Confluent.Kafka.IntegrationTests
             //   2. tasks should complete immediately.
             using (var producer = new ProducerBuilder<byte[], byte[]>(producerConfig).Build())
             {
-                Assert.Throws<InvalidOperationException>(() => producer.BeginProduce(
+                Assert.Throws<InvalidOperationException>(() => producer.Produce(
                     singlePartitionTopic,
                     new Message<byte[], byte[]> { Key = TestKey, Value = TestValue },
                     (DeliveryReport<byte[], byte[]> dr) => Console.WriteLine("should not print")));
 
-                Assert.Throws<InvalidOperationException>(() => producer.BeginProduce(
+                Assert.Throws<InvalidOperationException>(() => producer.Produce(
                     new TopicPartition(singlePartitionTopic, 0),
                     new Message<byte[], byte[]> { Key = TestKey, Value = TestValue },
                     (DeliveryReport<byte[], byte[]> dr) => Console.WriteLine("should not print")));
 
-                producer.BeginProduce(
+                producer.Produce(
                     new TopicPartition(singlePartitionTopic, 0),
                     new Message<byte[], byte[]> { Key = TestKey, Value = TestValue });
 
-                producer.BeginProduce(
+                producer.Produce(
                     singlePartitionTopic,
                     new Message<byte[], byte[]> { Key = TestKey, Value = TestValue });
 
-                producer.BeginProduce(
+                producer.Produce(
                     new TopicPartition(singlePartitionTopic, 0),
                     new Message<byte[], byte[]> { Key = TestKey, Value = TestValue });